==> Maskanalysis/0.1_Questionnaire data.Rmd <==
---
title: "questionnaire data"
author: "Katrina Li"
date: "05/01/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(data.table)
library(stringr)
```

# Extract age information & mask problem
From pre-experiment questionnaire: age, languages, mask problems
From post-experiment questionnaire: diffulctuly, distraction, whether experiment goes well

```{r cars}
path <- "questionnaire"
adultpath <- "questionnaire/adult questionnaire"
childpath <- "questionnaire/child questionnaire"
spreadsheetpath <- "questionnaire/experiment spreadsheet"
adultprefiles <- list.files(adultpath,pattern = "hykl.csv",full.names = T)
childprefiles <- list.files(childpath,pattern = "fsom.csv",full.names = T)
adultpostfiles <- list.files(adultpath,pattern = "tizj.csv",full.names = T)
childpostfiles <- list.files(childpath,pattern = "22wz.csv",full.names = T)

# A new method for reading all files at once: fread, the fatest options
# Single-line footer automatically disgarded
adultpre = do.call(rbind, lapply(adultprefiles, fread))
childpre = do.call(rbind, lapply(childprefiles, fread))
adultpost = do.call(rbind, lapply(adultpostfiles, fread))
childpost = do.call(rbind, lapply(childpostfiles, fread))

```

```{r clean data}
adultpost <- adultpost %>%
  select("Participant Public ID","Experiment Version","Difficulty","Distraction","response-4-1","response-4-2","response-4-3","response-4-4","response-4-text","response-5","Comments")

childpost <- childpost %>%
  select("Participant Public ID","Experiment Version","Difficulty","Distraction","response-4-1","response-4-2","response-4-3","response-4-4","response-4-text","response-5","Comments")

adult <- adultpre %>%
  select("Participant Public ID","Experiment Version","Participant OS","Participant Browser", "randomiser-t27m","date of birth-year","date of birth-month","sex","languages","response-6","response-6-text","mask problems","Difficulty","Tiring","more info on masks") %>%
  inner_join(adultpost, by = c("Participant Public ID","Experiment Version"))

child <- childpre %>%
  select("Participant Public ID","Experiment Version","Participant OS","Participant Browser", "randomiser-t27m","date of birth-year","date of birth-month","sex","languages","response-6","response-6-text","mask problems","Difficulty","Tiring","more info on masks") %>%
  inner_join(childpost, by = c("Participant Public ID","Experiment Version"))

# Remove the spacing in the variable so that filtering can work
names(adult)<-str_replace_all(names(adult), c(" " = ".","-"= "_"))
names(child)<-str_replace_all(names(child), c(" " = ".","-"= "_"))

```


```{r fix age problem}
colnames(adult)

adult %>%
  filter(date.of.birth_year == 0)

child %>%
  filter(date.of.birth_year == 0)

# 210725_c81 (Jackie) — 11;11
child$date.of.birth_year[child$Participant.Public.ID == "210725_c81"] <- 11
child$date.of.birth_month[child$Participant.Public.ID == "210725_c81"] <- 11
# 210725_c86 — 11;0
child$date.of.birth_year[child$Participant.Public.ID == "210725_c86"] <- 11
child$date.of.birth_month[child$Participant.Public.ID == "210725_c86"] <- 0
# 210725_c105 — 8;2
child$date.of.birth_year[child$Participant.Public.ID == "210725_c105"] <- 8
child$date.of.birth_month[child$Participant.Public.ID == "210725_c105"] <- 2
# 210818_c166 — 8;8
child$date.of.birth_year[child$Participant.Public.ID == "210818_c166"] <- 8
child$date.of.birth_month[child$Participant.Public.ID == "210818_c166"] <- 8
# 210725_c88 — 9;9
child$date.of.birth_year[child$Participant.Public.ID == "210725_c88"] <- 9
child$date.of.birth_month[child$Participant.Public.ID == "210725_c88"] <- 9
# 210725_c125 - 10;2
child$date.of.birth_year[child$Participant.Public.ID == "210725_c125"] <- 10
child$date.of.birth_month[child$Participant.Public.ID == "210725_c125"] <- 2
# 210725_c133 - 12;4 (17/4/2009)
child$date.of.birth_year[child$Participant.Public.ID == "210725_c133"] <- 12
child$date.of.birth_month[child$Participant.Public.ID == "210725_c133"] <- 4
# 210725_a86 — 31;4
adult$date.of.birth_year[adult$Participant.Public.ID == "210725_a86"] <- 31
adult$date.of.birth_month[adult$Participant.Public.ID == "210725_a86"] <- 4


adult %>%
  filter(date.of.birth_year == 0)

child %>%
  filter(date.of.birth_year == 0)
```

```{r}
write.csv(adult, "questionnaire/Qadult.csv", row.names = FALSE)
write.csv(child, "questionnaire/Qchild.csv", row.names = FALSE)
```
