==> Maskanalysis/0.2_Datacleaning.Rmd <==
---
title: "Data cleaning (first batch -- 17 files)"
author: "Katrina Li"
date: "28/10/2021"
output: html_document
---

```{r first batch, include=FALSE}
library(dplyr)
library(tidyr)
library(stringr)
library(readr)
```
update on 17/1/2022: start data point: **7853**

- from soundcheck_error, remove the points in discard (61), change 55 ACC=0 to ACC =1 and update their RT

- from soundcheck_RT, remove the points in discard (23), change 4 ACC=1 to ACC =0

ending in **7769** data points.


# Deal with trials modified in double check
```{r}
errorcheck <- read.delim("checkingnotes/soundcheck_error.txt", header = T) #317
keep <- subset(errorcheck, errorcheck$doublecheck == 1) #55
discard <- subset(errorcheck, errorcheck$doublecheck == 2) #61

RTcheck <- read.delim("checkingnotes/soundcheck_RT.txt", header = T) #98
RTdiscard <- subset(RTcheck, RTcheck$doublecheck == 2) #23
RTerror <- subset(RTcheck, RTcheck$doublecheck == 0) #4
```

```{r }
path <- "datafile_first_batch"
timep1files <- list.files(path,pattern = "p1.txt",full.names = T)
timep2files <- list.files(path,pattern = "p2.txt",full.names = T)

p1files <-tibble(File = timep1files) %>%
  extract(File, "parid","([0-9]{6}_[A-Za-z0-9]*time)",remove = FALSE) %>%
  mutate(Data = lapply(File, read.delim, sep = "\t")) %>%
  unnest(Data)%>%
  select(-File)%>%
  mutate(label = str_replace_all(label,"\\.","")) %>% # Replace extra '.' symbol at the end
  mutate(label = str_replace_all(label,"\\_$","")) %>%# Replace extra '_' symbol at the end
  mutate(label = str_replace_all(label,",$",""))

p2files <-tibble(File = timep2files) %>%
  extract(File, "parid","([0-9]{6}_[A-Za-z0-9]*time)",remove = FALSE) %>%
  mutate(Data = lapply(File, read.delim, sep = "\t",check.names = F)) %>%
  unnest(Data) %>%
  select(-File) %>%
  mutate(label = str_replace_all(label,"\\.",""))%>%
  mutate(label = str_replace_all(label,"\\_$","")) %>%
  mutate(label = str_replace_all(label,",$",""))

modifykeep <- keep %>% 
  left_join(p2files, by = "label") %>%
  drop_na() %>%
  mutate(
  duration_new = case_when(
    parid == "210719_a25time" ~ duration_raw,
    parid == "210719_c12time" ~ duration_raw,
    parid == "210719_a41time" ~ duration_raw - 0.8,
    parid == "210725_a83time" ~ duration_raw - 0.7,
    TRUE ~ duration_raw - 0.5
    )
  )%>%
  select(label, duration_new)

p2files <-p2files %>%
  left_join(modifykeep, by = "label") %>%
  mutate(
    duration2 = ifelse(is.na(duration_new), duration2, duration_new),
    note = ifelse(is.na(duration_new), note, sub("0","1",note))
  ) %>%
  select(-duration_new)

stimulim <- read.csv("datafile_first_batch/stimulim.csv",header = T)

#c29timep2, a120timep1&a120 timep2 have encoding problems. 
# Solution: open the file in Numbers, copy it to excel, and save as tab-delimited file txt

# Other possible errors: by using 'separte' failes below, you will be able to see 
combine <- p1files %>% 
  pivot_wider(
    names_from = timetype,
    values_from = timepoint,
  ) %>%
  mutate(
    duration1 = rstart-b
  ) %>%
  inner_join(p2files, by = c("parid","label")) %>% #anti_join could see which files are not matched
  mutate(
    beep_response = duration1 + duration2,
    original_label = label
  ) %>%
  separate(note, into=c("accuracy", "note"), sep="(?<=[0-1])_") %>%
  separate(label, into=c("scheduleID", "sentence"), sep = "_") %>% 
  separate(sentence, into = c('number','audio','video'), sep = "(a|v)(?=M|NM)") %>%
  mutate(
    note = str_replace_all(note,"^_",""), # Replace extra '_' symbol in the note
    video = str_replace_all(video, "[a-z]$","")
  )%>%
  mutate(
    audio = paste0("a",audio),
    video = paste0("v",video),
    sentence_label = paste(number, audio, sep="_")
  ) %>%
  left_join(stimulim, by = c("sentence_label","number")) %>%
  mutate(
    react1 = (beep_response - length1)*1000,
    react2 = (beep_response - length2)*1000,
    type = paste0(audio,video),
    context = str_sub(number,-1,-1),
    parid = str_sub(parid, 1, nchar(parid)-4),
    group = str_extract(parid,"(?<=_)[a,c]"),
    batch = 1
    #group = recode(group,a="adult",c="child")
  )  %>%
  select(parid, scheduleID, react1, react2,accuracy, note, number, correctans,audio, video, context, group, type, beep_response, length1, length2,batch,original_label) 

#write.csv(combine,"datafile/firstbatch.csv",row.names = FALSE)

# check group balanece
combine[!duplicated(combine$parid),] %>%
  group_by(group) %>%
  summarise(n = n())

# check if there are left-overs in merge
p1.wide <- p1files %>% 
  pivot_wider(
    names_from = timetype,
    values_from = timepoint,
  ) %>%
  mutate(
    duration1 = rstart-b
  ) 
#in timep1 but no match in timep2
p1.wide %>%
  anti_join(p2files, by = c("parid","label"))
#a120 has a wrong symbol added to video condition. saved it!

#in timep2 but no match in timep1
p2files %>%
  anti_join(p1.wide, by = c("parid","label"))
```

```{r second batch}
path_2 <- "datafile_second_batch"
timep1files_2 <- list.files(path_2,pattern = "p1.txt",full.names = T)
timep2files_2 <- list.files(path_2,pattern = "p2.txt",full.names = T)

p1files_2 <-tibble(File = timep1files_2) %>%
  extract(File, "parid","([0-9]{6}_[A-Za-z0-9]*time)",remove = FALSE) %>%
  mutate(Data = lapply(File, read.delim, sep = "\t")) %>%
  unnest(Data)%>%
  select(-File)%>%
  mutate(label = str_replace_all(label,"\\.$","")) %>% # Replace extra '.' symbol at the end
  mutate(label = str_replace_all(label,"\\_$","")) %>% # Replace extra '_' symbol at the end
  mutate(label = str_replace_all(label,",$","")) # Replace extra ',' symbol at the end

p2files_2 <-tibble(File = timep2files_2) %>%
  extract(File, "parid","([0-9]{6}_[A-Za-z0-9]*time)",remove = FALSE) %>%
  mutate(Data = lapply(File, read.delim, sep = "\t",check.names = F)) %>%
  unnest(Data) %>%
  select(-File) %>%
  mutate(label = str_replace_all(label,"\\.",""))%>%
  mutate(label = str_replace_all(label,"\\_$","")) %>%
  mutate(label = str_replace_all(label,",$",""))

modifykeep_2 <- keep %>% 
  left_join(p2files_2, by = "label") %>%
  drop_na() %>%
  mutate(
  duration_new = case_when(
    parid == "210719_a25time" ~ duration_raw,
    parid == "210719_c12time" ~ duration_raw,
    parid == "210719_a41time" ~ duration_raw - 0.8,
    parid == "210725_a83time" ~ duration_raw - 0.7,
    TRUE ~ duration_raw - 0.5
    )
  ) %>%
  select(label, duration_new)

p2files_2 <-p2files_2 %>%
  left_join(modifykeep_2, by = "label") %>%
  mutate(
    duration2 = ifelse(is.na(duration_new), duration2, duration_new),
    note = ifelse(is.na(duration_new), note, sub("0","1",note))
  ) %>%
  select(-duration_new)

stimulim <- read.csv("datafile_first_batch/stimulim.csv",header = T)

#c29timep2, a120timep1&a120 timep2 have encoding problems. 
# Solution: open the file in Numbers, copy it to excel, and save as tab-delimited file txt

# Other possible errors: by using 'separte' failes below, you will be able to see 
combine_2 <- p1files_2 %>% 
  pivot_wider(
    names_from = timetype,
    values_from = timepoint,
  ) %>%
  mutate(
    duration1 = rstart-b,
    
  ) %>%
  inner_join(p2files_2, by = c("parid","label")) %>% #anti_join could see which files are not matched
  mutate(
    beep_response = duration1 + duration2,
    original_label = label
  ) %>%
  separate(note, into=c("accuracy", "note"), sep="(?<=[0-1])_") %>%
  separate(label, into=c("scheduleID", "sentence"), sep = "_") %>% 
  separate(sentence, into = c('number','audio','video'), sep = "(a|v)(?=M|NM)") %>%
  mutate(
    note = str_replace_all(note,"^_",""), # Replace extra '_' symbol in the note
    video = str_replace_all(video, "[a-z]$","")
  )%>%
  mutate(
    audio = paste0("a",audio),
    video = paste0("v",video),
    sentence_label = paste(number, audio, sep="_")
  ) %>%
  left_join(stimulim, by = c("sentence_label","number")) %>%
  mutate(
    react1 = (beep_response - length1)*1000,
    react2 = (beep_response - length2)*1000,
    type = paste0(audio,video),
    context = str_sub(number,-1,-1),
    parid = str_sub(parid, 1, nchar(parid)-4),
    group = str_extract(parid,"(?<=_)[a,c]"),
    batch = 2
    #group = recode(group,a="adult",c="child")
  )  %>%
  select(parid, scheduleID, react1, react2,accuracy, note, number, correctans,audio, video, context, group, type, beep_response, length1, length2,batch,original_label) 

write.csv(combine_2,"datafile/secondbatch.csv",row.names = FALSE)

# check group balanece
combine_2[!duplicated(combine_2$parid),] %>%
  group_by(group) %>%
  summarize(n = n())

# check if there are left-overs in merge
p1.wide_2 <- p1files_2 %>% 
  pivot_wider(
    names_from = timetype,
    values_from = timepoint,
  ) %>%
  mutate(
    duration1 = rstart-b
  ) 

#in timep1 but no match in timep2
p1.wide_2 %>%
  anti_join(p2files_2, by = c("parid","label"))
#a10timep1 a '.' replaces the conditions

#in timep2 but no match in timep1
p2files_2 %>%
  anti_join(p1.wide_2, by = c("parid","label"))
```

r full data inspection

```{r}
data1 <- read.csv("datafile/firstbatch.csv",header = T)
data2 <- read.csv("datafile/secondbatch.csv", header = T)
fulldata <- rbind(data1,data2)

fulldata <- fulldata %>%
    mutate(
    videoname = paste(number, audio, video, sep="_"),
    videoname = paste(videoname,".mp4",sep = ""),
    accuracy = ifelse(original_label %in% RTerror$label, 0, accuracy),
    #accuracy = ifelse(original_label %in% keep$label, 1, accuracy) #changed in p2files already!
    )

table(fulldata$accuracy)

# first generate accuracy2 variable, then remove

# check NA values in the table?
fulldata[is.na(fulldata$react1),]
fulldata[is.na(fulldata$react2),]
fulldata[rowSums(is.na(fulldata)) >0,]
fulldata[!complete.cases(fulldata),]

# check if the audio/video/context condition matches - here match for sure, more problematic before merging p1 files and p2 files
table(fulldata$audio, fulldata$video, fulldata$context)

```

# Delete trials noted in 'Missing trials' & '2' in doublecheck files
```{r}
mis <- read.csv("datafile/missing_trials.csv", header = T)
mis <- subset(mis, mis$Missing !="NONE")

fulldata <- fulldata %>% #7873
  filter(!original_label %in% mis$Missing) %>% #7853
  filter(!original_label %in% discard$label) %>% #7792
  filter(!original_label %in% RTdiscard$label)

# check if some of the missingv trials noted in the spreadsheet are included inthe dataset
fulldata %>%
  filter(parid == "210723_a78" & number =="113L")

```



# Trial order & Questionnaire file

```{r}
trial_adult <- read.csv("questionnaire/adult_trial.csv", header = T)
trial_child <- read.csv("questionnaire/child_trial.csv", header = T)
trial <- rbind(trial_adult, trial_child)
trial <- trial %>%
  group_by(Participant.Public.ID,display,randomise_blocks) %>%
  mutate(
    rowmean = round(mean(row_number)),
  ) %>%
  ungroup %>%
  group_by(Participant.Public.ID) %>%
  mutate(
    block = case_when(
      rowmean == nth(unique(rowmean),1) ~ 0,
      rowmean == nth(unique(rowmean),2) ~1,
      rowmean == nth(unique(rowmean),3) ~2,
      rowmean == nth(unique(rowmean),4) ~3,
      rowmean == nth(unique(rowmean),5) ~4
    )
  )

#trial$block <- match(trial$rowmean, unique(trial$rowmean)) -1

# check if block number is correct
trial %>%
  group_by(Participant.Public.ID) %>%
  summarise(min = min(block), max = max(block), mean = mean(block), n = n())
# check if there is NA in trial (because of the way we name the variables)
trial %>%
  filter(is.na(block))
range(trial$block)

Qadult <- read.csv("questionnaire/Qadult.csv", header = T)
Qchild <- read.csv("questionnaire/Qchild.csv", header = T)
Qall <- rbind(Qadult, Qchild)

mergedata <- fulldata %>%
  left_join(trial, by = c("parid" = "Participant.Public.ID","videoname" = "videoname")) %>%
  left_join(Qall, by = c("parid"="Participant.Public.ID"))
```

# Basic Examination
```{r}
# Check the merge is sucessful
fulldata %>%
  anti_join(trial, by = c("parid" = "Participant.Public.ID","videoname" = "videoname")) 

# checking the number of NA for each column
# trial doesn't have weird single NA value
# Qall doesn't have weird single NA value
# full doesn't have weird single NA value
cbind(
   lapply(
     lapply(mergedata, is.na)
     , sum)
   )
mergedata[!complete.cases(mergedata),]

# Some variables has weird one NA where they shouldn't be
# Two sources of problem: adult, child questionnaire not exporting correctly; one strange data point in the check file
#210723_a79 full data has 134H_aNM_vM.mp4 not in the full data file!
# Go back and remove two Qs from the questionnaire: languages.text & mask.problems.text

# CHeck the range of block and trial, max.trial should be 120
# It is just possible that not all of the participants go over all stimuli
mergedata %>%
  group_by(parid) %>%
  summarise(max = max(row_number), n = n())

trial %>%
  group_by(Participant.Public.ID) %>%
  summarise(max = max(row_number), n = n())

#210725_a120 has 93LaMvNM as 121th -129 rows! two 25L!!!
#full data has alright
#it's trial's data fault - just two links - remove duplicates in python

trial %>%
  filter(Participant.Public.ID == "210725_a120") %>%
  filter(videoname == "25L_aM_vM.mp4") %>%
  select(filename)

#check ages are there
mergedata %>%
  filter(date.of.birth_year == 0)

mergedata %>%
  group_by(parid) %>%
  summarise(min = min(block), max = max(block), mean = mean(block), n = n())
range(trial$block)
# 210725_a156 9 blocks?

```


# Write the file
```{r read and generate full data}
nrow(mergedata)
colnames(mergedata)
write.csv(mergedata,"fulldata.csv",row.names = FALSE)
```
